Show CMs a list of related reports that they are allowed to see
